Refactor corrector interfaces for multiple shooting integration
- Updated the `_MultipleShootingCorrectorOrbitInterface` to enhance the handling of multiple shooting problems, including improved initialization of patches and residual computation.
- Refactored imports in `__init__.py` and `interfaces.py` for better organization and clarity.
- Removed the deprecated `_MultipleShootingCorrectorOrbitInterface` class from `interfaces_ms.py`, consolidating functionality into the main interface.
- Enhanced documentation for methods related to patch initialization, residuals, and Jacobians to improve usability and understanding of the multiple shooting process.
